pyproject.toml
[project]
name = "made-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.18",
    "python-dotenv>=1.1.0",
]
